refactor shared code
